Split operations test
